backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.12.0
dashscope==1.17.0
httpx==0.26.0
redis==5.0.1
chromadb==0.4.22
minio==7.2.3
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.2.0
numpy==1.26.3
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
sentence-transformers==2.3.1
openai-whisper==20231117
websockets==12.0
python-engineio==4.8.0
python-socketio==5.10.0
